+++ src/main/ts/components/Editor.tsx
@@ -12,7 +12,9 @@ import { ScriptLoader } from '../ScriptLoader';
 import { getTinymce } from '../TinyMCE';
 import { isFunction, isTextareaOrInput, mergePlugins, uuid, configHandlers, isBeforeInputEventAvailable, isInDoc, setMode } from '../Utils';
 import { EditorPropTypes, IEditorPropTypes } from './EditorPropTypes';
-import { Bookmark, Editor as TinyMCEEditor, EditorEvent, RawEditorSettings } from 'tinymce';
+import { Bookmark, Editor as TinyMCEEditor, EditorEvent, TinyMCE } from 'tinymce';
+
+type EditorOptions = Parameters<TinyMCE['init']>[0];
 
 export interface IProps {
   apiKey: string;
@@ -21,13 +23,11 @@ export interface IProps {
   initialValue: string;
   onEditorChange: (a: string, editor: TinyMCEEditor) => void;
   value: string;
-  init: RawEditorSettings & { selector?: undefined; target?: undefined };
-  /** @deprecated use `editor.getContent({format: 'text' })` in `onEditorChange` prop instead */
-  outputFormat: 'html' | 'text';
+  init: EditorOptions & { selector?: undefined; target?: undefined };
   tagName: string;
   cloudChannel: string;
-  plugins: NonNullable<RawEditorSettings['plugins']>;
-  toolbar: NonNullable<RawEditorSettings['toolbar']>;
+  plugins: NonNullable<EditorOptions['plugins']>;
+  toolbar: NonNullable<EditorOptions['toolbar']>;
   disabled: boolean;
   textareaName: string;
   tinymceScriptSrc: string;
@@ -48,7 +48,7 @@ export class Editor extends React.Component<IAllProps> {
   public static propTypes: IEditorPropTypes = EditorPropTypes;
 
   public static defaultProps: Partial<IAllProps> = {
-    cloudChannel: '5'
+    cloudChannel: '6'
   };
 
   public editor?: TinyMCEEditor;
@@ -271,9 +271,7 @@ export class Editor extends React.Component<IAllProps> {
       if (newContent !== this.currentContent) {
         this.currentContent = newContent;
         if (isFunction(this.props.onEditorChange)) {
-          const format = this.props.outputFormat;
-          const out = format === 'html' ? newContent : editor.getContent({ format });
-          this.props.onEditorChange(out, editor);
+          this.props.onEditorChange(newContent, editor);
         }
       }
     }
@@ -312,7 +310,7 @@ export class Editor extends React.Component<IAllProps> {
       throw new Error('tinymce should have been loaded into global scope');
     }
 
-    const finalInit: RawEditorSettings = {
+    const finalInit: EditorOptions = {
       ...this.props.init,
       selector: undefined,
       target,